refactor: apply nice-to-have review findings to Maple GraphQL indexer
- Include a bounded response-body snippet in 429/5xx client errors so
  retry-exhausted failures carry the upstream diagnostic
- Drop unused MapleActiveLoan.PoolName/PoolAssetSymbol/PoolAssetDecimals
  DTO fields and trim the loans query to fundingPool { id }; pool details
  flow from GetPools in the same cycle
- Add entity.MapleLoanTypeOTL const for the loan_type discriminator
- Deduplicate the four pgx.Batch RETURNING-id scan loops into a generic
  collectBatchIDs helper in the shared postgres helpers file
- Deduplicate ~20 repeated big.Int validation checks across the five
  maple entities into requireNonNegBigInt(IfSet) helpers
- Widen ServiceConfig.ChainID to int64 (matches sibling services),
  converting once at the cmd boundary instead of five casts in the service
- Clarify maple.sync.rows.written semantics: rows attempted, not inserted
  (ON CONFLICT dedup on same-build retries may insert fewer)
- Comment the no-op DO UPDATE trick that makes RETURNING yield existing ids
- Tests: per-parser malformed-value coverage for pools/strategies/globals,
  multi-chunk SavePoolStates integration test (batchSize=2), body-snippet
  assertion on the HTTP 500 retry test

+++ stl-verify/cmd/cronjobs/maple-graphql-indexer/main.go
@@ -84,7 +84,7 @@ func setupRunner(ctx context.Context, deps temporal.Dependencies) (temporal.Runn
 	}
 
 	service, err := maple_graphql_indexer.NewService(maple_graphql_indexer.ServiceConfig{
-		ChainID: chainID,
+		ChainID: int64(chainID),
 		Logger:  deps.Logger,
 	}, client, repo, txManager, telemetry)
 	if err != nil {
